+++ arc/utils/hooks/limiters.py
@@ -1,25 +1,15 @@
 from __future__ import annotations
 
-import asyncio
-import sys
-import time
-import traceback
 import typing as t
-from collections import deque
 
-import attr
-
-from arc.abc.hookable import HookResult
 from arc.abc.limiter import LimiterProto
+from arc.context.base import Context
 from arc.errors import UnderCooldownError
 from arc.internal.types import ClientT
-
-if t.TYPE_CHECKING:
-    from arc.context.base import Context
-
+from arc.utils.ratelimiter import RateLimiter, RateLimiterExhaustedError
 
 __all__ = (
-    "RateLimiter",
+    "LimiterHook",
     "global_limiter",
     "guild_limiter",
     "channel_limiter",
@@ -29,89 +19,8 @@ __all__ = (
 )
 
 
-@attr.define(slots=True, kw_only=True)
-class _Bucket(t.Generic[ClientT]):
-    """Handles the ratelimiting of a single item. (E.g. a single user or a channel)."""
-
-    key: str
-    """The key of the bucket."""
-
-    reset_at: float
-    """The time at which the bucket resets."""
-
-    limiter: RateLimiter[ClientT]
-    """The limiter this bucket belongs to."""
-
-    _remaining: int = attr.field(alias="remaining")
-    """The amount of requests remaining until the bucket is exhausted."""
-
-    _queue: deque[asyncio.Event] = attr.field(factory=deque, init=False)
-    """A list of events to set as the iter task proceeds."""
-
-    _task: asyncio.Task[None] | None = attr.field(default=None, init=False)
-    """The task that is currently iterating over the queue."""
-
-    @classmethod
-    def for_limiter(cls, key: str, limiter: RateLimiter[ClientT]) -> _Bucket[ClientT]:
-        """Create a new bucket for a RateLimiter."""
-        return cls(key=key, limiter=limiter, reset_at=time.monotonic() + limiter.period, remaining=limiter.limit)
-
-    @property
-    def remaining(self) -> int:
-        """The amount of requests remaining until the bucket is exhausted."""
-        if self.reset_at <= time.monotonic():
-            self.reset()
-        return self._remaining
-
-    @remaining.setter
-    def remaining(self, value: int) -> None:
-        self._remaining = value
-
-    @property
-    def is_exhausted(self) -> bool:
-        """Return a boolean determining if the bucket is exhausted."""
-        return self.remaining <= 0 and self.reset_at > time.monotonic()
-
-    @property
-    def is_stale(self) -> bool:
-        """Return a boolean determining if the bucket is stale.
-        If a bucket is stale, it is no longer in use and can be purged.
-        """
-        return not self._queue and self.remaining == self.limiter.limit and (self._task is None or self._task.done())
-
-    def start_queue(self) -> None:
-        """Start the queue of a bucket.
-        This will start setting events in the queue until the bucket is ratelimited.
-        """
-        if self._task is None or self._task.done():
-            self._task = asyncio.create_task(self._iter_queue())
-
-    def reset(self) -> None:
-        """Reset the bucket."""
-        self.reset_at = time.monotonic() + self.limiter.period
-        self._remaining = self.limiter.limit
-
-    async def _iter_queue(self) -> None:
-        """Iterate over the queue and set events until exhausted."""
-        try:
-            while self._queue:
-                if self.remaining <= 0 and self.reset_at > time.monotonic():
-                    # Sleep until ratelimit expires
-                    await asyncio.sleep(self.reset_at - time.monotonic())
-                    self.reset()
-
-                # Set events while not ratelimited
-                while self.remaining > 0 and self._queue:
-                    self._queue.popleft().set()
-                    self._remaining -= 1
-
-        except Exception as e:
-            print(f"Task Exception was never retrieved: {e}", file=sys.stderr)
-            print(traceback.format_exc(), file=sys.stderr)
-
-
-class RateLimiter(LimiterProto[ClientT]):
-    """The default implementation of a ratelimiter.
+class LimiterHook(RateLimiter[Context[ClientT]], LimiterProto[ClientT]):
+    """The default implementation of a ratelimiter that can be used as a hook.
 
     Parameters
     ----------
@@ -121,56 +30,18 @@ class RateLimiter(LimiterProto[ClientT]):
         The amount of requests allowed in a bucket.
     get_key_with : t.Callable[[Context[t.Any]], str]
         A callable that returns a key for the ratelimiter bucket.
+
+    See Also
+    --------
+    - [`global_limiter`][arc.utils.hooks.limiters.global_limiter]
+    - [`guild_limiter`][arc.utils.hooks.limiters.guild_limiter]
+    - [`channel_limiter`][arc.utils.hooks.limiters.channel_limiter]
+    - [`user_limiter`][arc.utils.hooks.limiters.user_limiter]
+    - [`member_limiter`][arc.utils.hooks.limiters.member_limiter]
+    - [`custom_limiter`][arc.utils.hooks.limiters.custom_limiter]
     """
 
-    __slots__ = ("period", "limit", "_buckets", "_get_key")
-
-    def __init__(self, period: float, limit: int, *, get_key_with: t.Callable[[Context[t.Any]], str]) -> None:
-        self.period: float = period
-        self.limit: int = limit
-        self._buckets: t.Dict[str, _Bucket[ClientT]] = {}
-        self._get_key: t.Callable[[Context[t.Any]], str] = get_key_with
-        self._gc_task: asyncio.Task[None] | None = None
-
-    def get_key(self, ctx: Context[t.Any]) -> str:
-        """Get key for ratelimiter bucket."""
-        return self._get_key(ctx)
-
-    def is_rate_limited(self, ctx: Context[t.Any]) -> bool:
-        """Returns a boolean determining if the ratelimiter is ratelimited or not.
-
-        Parameters
-        ----------
-        ctx : Context[t.Any]
-            The context to evaluate the ratelimit under.
-
-        Returns
-        -------
-        bool
-            A boolean determining if the ratelimiter is ratelimited or not.
-        """
-        now = time.monotonic()
-
-        if data := self._buckets.get(self.get_key(ctx)):
-            if data.reset_at <= now:
-                return False
-            return data._remaining <= 0
-        return False
-
-    def _start_gc(self) -> None:
-        """Start the garbage collector task if one is not running."""
-        if self._gc_task is None or self._gc_task.done():
-            self._gc_task = asyncio.create_task(self._gc())
-
-    async def _gc(self) -> None:
-        """Purge stale buckets."""
-        while self._buckets:
-            await asyncio.sleep(self.period + 1.0)
-            for bucket in list(self._buckets.values()):
-                if bucket.is_stale:
-                    del self._buckets[bucket.key]
-
-    async def acquire(self, ctx: Context[t.Any], *, wait: bool = True) -> None:
+    async def acquire(self, ctx: Context[ClientT], *, wait: bool = True) -> None:
         """Acquire a bucket, block execution if ratelimited and wait is True.
 
         Parameters
@@ -186,54 +57,15 @@ class RateLimiter(LimiterProto[ClientT]):
         UnderCooldownError
             If the ratelimiter is ratelimited and wait is False.
         """
-        event = asyncio.Event()
-
-        key = self.get_key(ctx)
-        # Get existing or insert new bucket
-        bucket = self._buckets.setdefault(key, _Bucket.for_limiter(key, self))
-
-        if bucket.is_exhausted and not wait:
+        try:
+            return await super().acquire(ctx, wait=wait)
+        except RateLimiterExhaustedError as exc:
             raise UnderCooldownError(
-                self,
-                bucket.reset_at - time.monotonic(),
-                f"Ratelimited for {bucket.reset_at - time.monotonic()} seconds.",
-            )
-
-        bucket._queue.append(event)
-        bucket.start_queue()
-        self._start_gc()
-
-        if wait:
-            await event.wait()
-
-    async def __call__(self, ctx: Context[t.Any]) -> HookResult:
-        """Acquire a ratelimit, fail if ratelimited.
-
-        Parameters
-        ----------
-        ctx : Context[t.Any]
-            The context to evaluate the ratelimit under.
-
-        Returns
-        -------
-        HookResult
-            A hook result to conform to the limiter protocol.
-
-        Raises
-        ------
-        UnderCooldownError
-            If the ratelimiter is ratelimited.
-        """
-        await self.acquire(ctx, wait=False)
-        return HookResult()
-
-    def reset(self, ctx: Context[t.Any]) -> None:
-        """Reset the ratelimit for a given context."""
-        if bucket := self._buckets.get(self.get_key(ctx)):
-            bucket.reset()
+                self, exc.retry_after, f"Command is under cooldown for '{exc.retry_after}' seconds."
+            ) from exc
 
 
-def global_limiter(period: float, limit: int) -> RateLimiter[t.Any]:
+def global_limiter(period: float, limit: int) -> LimiterHook[t.Any]:
     """Create a global ratelimiter.
 
     This ratelimiter is shared across all invocation contexts.
@@ -244,11 +76,17 @@ def global_limiter(period: float, limit: int) -> RateLimiter[t.Any]:
         The period, in seconds, after which the bucket resets.
     limit : int
         The amount of requests allowed in a bucket.
+
+    Usage
+    -----
+    ```py
+    @arc.with_hook(arc.global_limiter(5.0, 1))
+    ```
     """
-    return RateLimiter(period, limit, get_key_with=lambda _: "amongus")
+    return LimiterHook(period, limit, get_key_with=lambda _: "amongus")
 
 
-def guild_limiter(period: float, limit: int) -> RateLimiter[t.Any]:
+def guild_limiter(period: float, limit: int) -> LimiterHook[t.Any]:
     """Create a guild ratelimiter.
 
     This ratelimiter is shared across all contexts in a guild.
@@ -259,11 +97,17 @@ def guild_limiter(period: float, limit: int) -> RateLimiter[t.Any]:
         The period, in seconds, after which the bucket resets.
     limit : int
         The amount of requests allowed in a bucket.
+
+    Usage
+    -----
+    ```py
+    @arc.with_hook(arc.guild_limiter(5.0, 1))
+    ```
     """
-    return RateLimiter(period, limit, get_key_with=lambda ctx: str(ctx.guild_id))
+    return LimiterHook(period, limit, get_key_with=lambda ctx: str(ctx.guild_id))
 
 
-def channel_limiter(period: float, limit: int) -> RateLimiter[t.Any]:
+def channel_limiter(period: float, limit: int) -> LimiterHook[t.Any]:
     """Create a channel ratelimiter.
 
     This ratelimiter is shared across all contexts in a channel.
@@ -274,11 +118,17 @@ def channel_limiter(period: float, limit: int) -> RateLimiter[t.Any]:
         The period, in seconds, after which the bucket resets.
     limit : int
         The amount of requests allowed in a bucket.
+
+    Usage
+    -----
+    ```py
+    @arc.with_hook(arc.channel_limiter(5.0, 1))
+    ```
     """
-    return RateLimiter(period, limit, get_key_with=lambda ctx: str(ctx.channel_id))
+    return LimiterHook(period, limit, get_key_with=lambda ctx: str(ctx.channel_id))
 
 
-def user_limiter(period: float, limit: int) -> RateLimiter[t.Any]:
+def user_limiter(period: float, limit: int) -> LimiterHook[t.Any]:
     """Create a user ratelimiter.
 
     This ratelimiter is shared across all contexts by a user.
@@ -289,11 +139,17 @@ def user_limiter(period: float, limit: int) -> RateLimiter[t.Any]:
         The period, in seconds, after which the bucket resets.
     limit : int
         The amount of requests allowed in a bucket.
+
+    Usage
+    -----
+    ```py
+    @arc.with_hook(arc.user_limiter(5.0, 1))
+    ```
     """
-    return RateLimiter(period, limit, get_key_with=lambda ctx: str(ctx.author.id))
+    return LimiterHook(period, limit, get_key_with=lambda ctx: str(ctx.author.id))
 
 
-def member_limiter(period: float, limit: int) -> RateLimiter[t.Any]:
+def member_limiter(period: float, limit: int) -> LimiterHook[t.Any]:
     """Create a member ratelimiter.
 
     This ratelimiter is shared across all contexts by a member in a guild.
@@ -305,11 +161,17 @@ def member_limiter(period: float, limit: int) -> RateLimiter[t.Any]:
         The period, in seconds, after which the bucket resets.
     limit : int
         The amount of requests allowed in a bucket.
+
+    Usage
+    -----
+    ```py
+    @arc.with_hook(arc.member_limiter(5.0, 1))
+    ```
     """
-    return RateLimiter(period, limit, get_key_with=lambda ctx: f"{ctx.author.id}:{ctx.guild_id}")
+    return LimiterHook(period, limit, get_key_with=lambda ctx: f"{ctx.author.id}:{ctx.guild_id}")
 
 
-def custom_limiter(period: float, limit: int, get_key_with: t.Callable[[Context[t.Any]], str]) -> RateLimiter[t.Any]:
+def custom_limiter(period: float, limit: int, get_key_with: t.Callable[[Context[t.Any]], str]) -> LimiterHook[t.Any]:
     """Create a ratelimiter with a custom key extraction function.
 
     Parameters
@@ -322,8 +184,14 @@ def custom_limiter(period: float, limit: int, get_key_with: t.Callable[[Context[
         A callable that returns a key for the ratelimiter bucket. This key is used to identify the bucket.
         For instance, to create a ratelimiter that is shared across all contexts in a guild,
         you would use `lambda ctx: str(ctx.guild_id)`.
+
+    Usage
+    -----
+    ```py
+    @arc.with_hook(arc.custom_limiter(5.0, 1, lambda ctx: str(ctx.guild_id)))
+    ```
     """
-    return RateLimiter(period, limit, get_key_with=get_key_with)
+    return LimiterHook(period, limit, get_key_with=get_key_with)
 
 
 # MIT License