+++ docs/changelog.md
@@ -14,6 +14,8 @@ Here you can find all the changelogs for `hikari-arc`.
 - Add `Context.get_option` to access options outside the command callback in a type-safe manner.
 - Add `Client.walk_commands()` and `Plugin.walk_commands()` to iterate over all commands & subcommands of a given type in a type-safe manner.
 - Add `CallableCommandProto.display_name`, `SlashCommand.make_mention()`, `SlashSubCommand.make_mention()`.
+- Remove the `acquire()` method from `LimiterProto` to make it easier to implement custom limiters.
+- Split `arc.utils.hooks.RateLimiter` into `arc.utils.RateLimiter` and `arc.utils.hooks.LimiterHook`. This allows `arc.utils.RateLimiter` to be used independently of an arc context object.
 
 ## v0.5.0